% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attrakttv-app.R
\name{attrakttv}
\alias{attrakttv}
\title{Run attrakttv app}
\usage{
attrakttv(display.mode = "auto",
  launch.browser = getOption("shiny.launch.browser", interactive()))
}
\arguments{
\item{display.mode}{\code{auto} by default, can also be \code{showcase}.
See \link[shiny]{runApp}.}

\item{launch.browser}{Boolean, set \code{TRUE} to open the app in the browser.
See \link[shiny]{runApp}.}
}
\description{
This function starts the attrakttv shiny app.
}
\examples{
\dontrun{
attrakttv()
}
}
