% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_helpers.R
\name{cach_db_path}
\alias{cach_db_path}
\title{Get the path to the db file}
\usage{
cach_db_path(name = "tRakt.db")
}
\arguments{
\item{name}{Optional: Name of db file. Defaults to \code{tRakt.db}}
}
\value{
\code{character(1)}
}
\description{
Get the path to the db file
}
\examples{
cach_db_path()
}
