% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_helpers.R
\name{cache_add_data}
\alias{cache_add_data}
\title{Add data to some db table}
\usage{
cache_add_data(table_name, new_data, replace = FALSE, cache_db_con)
}
\arguments{
\item{table_name}{Name of the db table to add data to.}

\item{new_data}{The new data to add.}

\item{replace}{\code{logical [FALSE]} Delete (if exists) & replace data?}

\item{cache_db_con}{A connection, see \link{cache_db}}
}
\value{
Nothing
}
\description{
Add data to some db table
}
\examples{
\dontrun{
TRUE
}
}
