% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_helpers.R
\name{cache_add_poster}
\alias{cache_add_poster}
\title{Add a poster to the cache}
\usage{
cache_add_poster(show_id, replace = FALSE, cache_db_con)
}
\arguments{
\item{show_id}{A show id, in the database the \verb{trakt ID} is used.}

\item{replace}{\code{logical [FALSE]} Delete (if exists) & replace data?}

\item{cache_db_con}{A connection, see \link{cache_db}}
}
\value{
Nothing
}
\description{
Add a poster to the cache
}
