% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_helpers.R
\name{cache_add_show}
\alias{cache_add_show}
\title{Add a show to the cache}
\usage{
cache_add_show(
  show_query = NULL,
  show_id = NULL,
  replace = FALSE,
  cache_db_con
)
}
\arguments{
\item{show_query}{A search query, e.g. "Futurama".}

\item{show_id}{A show id, in the database the \verb{trakt ID} is used.}

\item{replace}{\code{logical [FALSE]} Delete (if exists) & replace data?}

\item{cache_db_con}{A connection, see \link{cache_db}}
}
\value{
\code{NULL} \emph{if} the \code{search_query} yields no result, the \code{show_id} otherwise.
}
\description{
Add a show to the cache
}
