% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_helpers.R
\name{cache_add_show_query}
\alias{cache_add_show_query}
\title{Add a show by query}
\usage{
cache_add_show_query(show_query, replace = FALSE, cache_db_con)
}
\arguments{
\item{show_query}{A search query, e.g. "Futurama".}

\item{replace}{\code{logical [FALSE]} Delete (if exists) & replace data?}

\item{cache_db_con}{A connection, see \link{cache_db}}
}
\description{
Just a workhorse behind \link{cache_add_show}.
}
\keyword{internal}
