% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_helpers.R
\name{cache_drop_old_rows}
\alias{cache_drop_old_rows}
\title{Drop old rows}
\usage{
cache_drop_old_rows(table_name, threshold_days = 7, cache_db_con)
}
\arguments{
\item{threshold_days}{\code{integer [7]}: Drop records older than that.}

\item{cache_db_con}{A connection, see \link{cache_db}}
}
\value{
Nothing
}
\description{
Drop old rows
}
