% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_helpers.R
\name{cache_update_episodes}
\alias{cache_update_episodes}
\title{Update outdated episodes (in bulk per show)}
\usage{
cache_update_episodes(criterion = "aired")
}
\arguments{
\item{criterion}{\code{character(1) ["aired"]} Which criterion to use for updating. Currently
only \code{"aired"} is implemented, which updates all seasons if there are episodes that
aired after \code{cache_date}}
}
\value{
Nothing
}
\description{
Update outdated episodes (in bulk per show)
}
\examples{
\dontrun{
cache_update_episodes()
}
}
