% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_helpers.R
\name{check_cache_table}
\alias{check_cache_table}
\title{Check if a table exists in db, if not, create it}
\usage{
check_cache_table(table_name, reference_table, cache_db_con)
}
\arguments{
\item{table_name}{A db table name, e.g. "shows".}

\item{reference_table}{A table used as a template for (re)creation of the db table.}

\item{cache_db_con}{A connection, see \link{cache_db}}
}
\value{
Nothing
}
\description{
Check if a table exists in db, if not, create it
}
\examples{
\dontrun{
TRUE
}
}
