% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_helpers.R
\name{convert_ids}
\alias{convert_ids}
\title{Convert back and forth between slugs and show_ids}
\usage{
convert_ids(show_id = NULL, slug = NULL, cache_db_con)
}
\arguments{
\item{show_id, slug}{A \code{show_id} or \code{slug}, of which \emph{one} must
be supplied to get the other.}

\item{cache_db_con}{A connection, see \link{cache_db}}
}
\value{
\code{character(1)}
}
\description{
Convert back and forth between slugs and show_ids
}
\examples{
convert_ids(slug = "futurama")
convert_ids(show_id = "614")
}
