% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_helpers.R
\name{country_label}
\alias{country_label}
\title{Label country codes}
\usage{
country_label(x)
}
\arguments{
\item{x}{A country code like \code{"us"}}
}
\value{
\code{character} of same length as \code{x}
}
\description{
Label country codes
}
