% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_helpers.R
\name{days_ago}
\alias{days_ago}
\title{Datetime of last week}
\usage{
days_ago(days = 7L, unix = TRUE)
}
\arguments{
\item{days}{\code{integer [7]}: How many days ago? Must be an integer or coercible to an integer.}

\item{unix}{\code{logical [TRUE]}: Return a unix timestamp (which the db understands).}
}
\value{
Either a \code{POSIXct} or \code{numeric}, see \code{unix}.
}
\description{
Datetime of last week
}
\examples{
days_ago(1)
days_ago(unix = FALSE)
}
