% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_helpers.R
\name{is_already_cached}
\alias{is_already_cached}
\title{Is the show in the cache already?}
\usage{
is_already_cached(table_name, show_id, cache_db_con)
}
\arguments{
\item{table_name}{A db table name, e.g. "shows".}

\item{show_id}{A show id, in the database the \code{trakt ID} is used.}

\item{cache_db_con}{A connection, see \link{cache_db}}
}
\value{
\code{logical(1)}, \code{TRUE} if the \code{show_id} is already present in the table.
}
\description{
Is the show in the cache already?
}
\examples{
\dontrun{
TRUE
}
}
