% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_helpers.R
\name{language_label}
\alias{language_label}
\title{Label language codes}
\usage{
language_label(x)
}
\arguments{
\item{x}{A language code like \code{"en"}}
}
\value{
\code{character} of same length as \code{x}
}
\description{
Label language codes
}
\examples{
language_label(c("us", "de", "lwaflaf"))
}
