% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_helpers.R
\name{make_hoverinfo}
\alias{make_hoverinfo}
\title{Append hovertext to episodes df}
\usage{
make_hoverinfo(episodes)
}
\arguments{
\item{episodes}{A tbl as it will arrive in \code{renderPlotly}}
}
\value{
The same tbl with a new column \code{hovertext}
}
\description{
Append hovertext to episodes df
}
