% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_helpers.R
\name{rating_label}
\alias{rating_label}
\title{Label user ratings}
\usage{
rating_label(x)
}
\arguments{
\item{x}{Numeric rating in \verb{[1, 10]}. Will be rounded to the nearest integer.}
}
\value{
\code{character} of same length as \code{x}.
}
\description{
Label user ratings
}
