% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_helpers.R
\name{unix_date}
\alias{unix_date}
\title{Convert numeric timestam to Date}
\usage{
unix_date(x)
}
\arguments{
\item{x}{A numeric timestamp in unix epoch.}
}
\value{
Object of class \code{Date}.
}
\description{
Note: Hardcoded UTC timezone and origin of \code{1970-01-01}
}
\examples{
x <- c(927079200, 958960800, 989805600, 1060567200, 1315533600, 1378346400)
unix_date(x)
}
