% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_helpers.R
\name{unix_datetime}
\alias{unix_datetime}
\title{Convert numeric timestamp to POSIXct}
\usage{
unix_datetime(x)
}
\arguments{
\item{x}{A numeric timestamp in unix epoch.}
}
\value{
Object of class \code{POSIXct}.
}
\description{
Convert numeric timestamp to POSIXct
}
