% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attrakttv-package.R
\docType{package}
\name{attrakttv-package}
\alias{attrakttv}
\alias{attrakttv-package}
\title{attrakttv: A Shiny Application for 'trakt.tv' TV Show Data}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Lukas Burk \email{lukas@tadaa-data.de} (0000-0001-7528-3795)

}
\keyword{internal}
