% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attrakttv-package.R
\docType{package}
\name{attrakttv-package}
\alias{attrakttv}
\alias{attrakttv-package}
\title{attrakttv: A Shiny Application for 'trakt.tv' TV Show Data}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jemus42/attrakttv}
  \item Report bugs at \url{https://github.com/jemus42/attrakttv/issues}
}

}
\author{
\strong{Maintainer}: Lukas Burk \email{lukas@tadaa-data.de} (\href{https://orcid.org/0000-0001-7528-3795}{ORCID})

}
\keyword{internal}
