% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_helpers.R
\name{cache_db}
\alias{cache_db}
\title{Make a connection to the db}
\usage{
cache_db(pool = TRUE)
}
\arguments{
\item{pool}{\code{logical(1) [TRUE]}: Return a \link{pool}.}
}
\value{
A \code{conn} \link{DBI} thingy \emph{or} \code{pool}.
}
\description{
Make a connection to the db
}
\examples{
\dontrun{
cache_db_con <- cache_db()

is_already_cached("shows", 1390, cache_db_con)
}
}
