% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_helpers.R
\name{cache_db_path}
\alias{cache_db_path}
\title{Get the path to the db file}
\usage{
cache_db_path(name = "tRakt.db")
}
\arguments{
\item{name}{Optional: Name of db file. Defaults to \code{tRakt.db}}
}
\value{
\code{character(1)}
}
\description{
The directory is set to \code{getOption("trakt_db_path", default = "~/db")}.
}
\examples{
cache_db_path()
}
