% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_helpers.R
\name{cache_db_path}
\alias{cache_db_path}
\title{Get the path to the db file}
\usage{
cache_db_path(name = "tRakt.db", verbose = TRUE)
}
\arguments{
\item{name}{\verb{["tRakt.db"]}: Name of db file.}

\item{verbose}{\verb{[TRUE]}: Log the db path to the console.}
}
\value{
Invisibly: \code{character(1)}: Full path to the db file.
}
\description{
The directory is set to \code{Sys.getenv("trakt_db_path")}, with a default location
derived from \code{rappdirs::user_data_dir()} if the variable is not set.
}
\details{
The default directories are:
\itemize{
\item macOS: \code{"/Users/<user>/Library/Application Support/attrakttv/<version>/tRakt.db"}
\item Linux \code{"/home/<user>/.local/share/attrakttv/<version>/tRakt.db"} (\verb{$"XDG_DATA_HOME"})
\item Windows \verb{""C:/Users/<user>/Local/attrakttv/attrakttv/<version>/tRakt.db""}
}
}
\examples{
cache_db_path()
}
