% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_helpers.R
\name{cleanup_show_summary}
\alias{cleanup_show_summary}
\title{Prepare a show info object for the database}
\usage{
cleanup_show_summary(show)
}
\arguments{
\item{show}{A show object as returned by \link[tRakt:trakt.shows.summary]{tRakt::trakt.shows.summary} or \link[tRakt:trakt.search]{tRakt::trakt.search}
with \code{extended = "full"}.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Prepare a show info object for the database
}
\examples{
tRakt::trakt.search("Futurama", type = "shows") \%>\%
  cleanup_show_summary()
}
