% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_helpers.R
\name{db_init}
\alias{db_init}
\title{Inititale database based on seed data}
\usage{
db_init(path = cache_db_path(), cache_db_con = cache_db(pool = FALSE))
}
\arguments{
\item{path}{Defaults to \code{cache_db_path()}.}

\item{cache_db_con}{Defaults to \code{cache_db(pool = FALSE)}}
}
\value{
Nothing
}
\description{
Creates tables in db at \code{path} if not existing yet.
}
