% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_helpers.R
\name{get_fanart_poster}
\alias{get_fanart_poster}
\title{Get a poster from fanart.tv}
\usage{
get_fanart_poster(tvdbid, api_key = Sys.getenv("fanarttv_api_key"))
}
\arguments{
\item{tvdbid}{THe tvdb id.}

\item{api_key}{Defaults to \code{Sys.getenv("fanarttv_api_key")}.}
}
\value{
A \code{character} with a poster url. If there's no result, return is \code{""} (\code{character(1)})
}
\description{
Get a poster from fanart.tv
}
\examples{
\dontrun{
get_fanart_poster()
}
}
